add Image controllers, erros, middlewares

+++ src/controllers/TestimonialsController.ts
@@ -1,43 +1,56 @@
-import { Request, Response } from "express";
+import { NextFunction, Request, Response } from "express";
 import { testimonialRepository } from "../repositories/testimonial";
 import { findOneUser, userRepository } from "../repositories/user";
 import Testimonial from "../models/Testimonial";
 import { manager } from "../config/data-source";
 import User from "../models/User";
+import NotFound from "../errors/NotFound";
 
 class TestimonialsControllers {
-
-	static async listAllTestimonials(_req: Request, res: Response) {
+	static async listAllTestimonials(
+		_req: Request,
+		res: Response,
+		next: NextFunction
+	) {
 		try {
 			const users: User[] = await userRepository
 				.find({
 					relations: {
 						testemonials: true,
-						photo: true
+						photo: true,
 					},
 				})
 				.then((users) =>
 					users.filter((user) => user.testemonials.length !== 0)
 				);
-			return res.status(200).json(users);
+
+			return users.length > 0
+				? res.status(200).json(users)
+				: next(new NotFound());
 		} catch (error) {
-			return res.status(500).json(error.message);
+			next(error);
 		}
 	}
 
-	static async listAllTestimonialsByUser(req: Request, res: Response) {
+	static async listAllTestimonialsByUser(
+		req: Request,
+		res: Response,
+		next: NextFunction
+	) {
 		const { id }: any = req.params;
 
 		try {
-			const users: User = await userRepository.findOne({
+			const user: User = await userRepository.findOne({
 				where: { id: id },
-				relations: { 
+				relations: {
 					testemonials: true,
-					photo: true
+					photo: true,
 				},
 			});
 
-			return res.status(200).json(users);
+			return user
+				? res.status(200).json(user)
+				: next(new NotFound(`usuario com o ${id} não encontrado`));
 		} catch (error) {
 			return res.status(500).json(error.message);
 		}
@@ -45,15 +58,14 @@ class TestimonialsControllers {
 
 	static async addTestimonial(req: Request, res: Response) {
 		const { userId, text }: any = req.body;
-		const message: string = 'depoimento adicionado com sucesso!';
+		const message: string = "depoimento adicionado com sucesso!";
 
 		try {
 			const user: User = await findOneUser(userId, true);
 			const newTestimonial = new Testimonial(text, user);
 			await testimonialRepository.save(newTestimonial);
-			return res
-				.status(201)
-				.json({ message: message });
+
+			return res.status(201).json({ message: message });
 		} catch (error) {
 			return res.status(500).json(error.message);
 		}
@@ -62,17 +74,19 @@ class TestimonialsControllers {
 	static async updateTestimonial(req: Request, res: Response) {
 		const { id }: any = req.params;
 		const { text }: any = req.body;
-		const message: string = 'depoimento atualizado com sucesso!';
+		const message: string = "depoimento atualizado com sucesso!";
 
 		try {
 			const testimonial: Testimonial =
 				await testimonialRepository.findOneBy({ id });
-			testimonial.text = text;
 
-			await testimonialRepository.save(testimonial);
-			return res
-				.status(201)
-				.json({ message: message });
+			if (testimonial) {
+				testimonial.text = text;
+				await testimonialRepository.save(testimonial);
+				return res.status(201).json({ message: message });
+			} else {
+				return res.status(404).send({});
+			}
 		} catch (error) {
 			return res.status(500).json(error.message);
 		}
@@ -80,27 +94,33 @@ class TestimonialsControllers {
 
 	static async deleteTestimonial(req: Request, res: Response) {
 		const { id }: any = req.params;
-		const message: string ='depoimento com o id ${id} foi deletado';
+		const message: string = "depoimento com o id ${id} foi deletado";
 
 		try {
 			const testimonial: Testimonial =
 				await testimonialRepository.findOneBy({ id });
-			await testimonialRepository.remove(testimonial);
-			return res
-				.status(200)
-				.json({ message: message });
+			if (testimonial) {
+				await testimonialRepository.remove(testimonial);
+				return res.status(200).json({ message: message });
+			} else {
+				return res.status(404).send({});
+			}
 		} catch (error) {
 			return res.status(500).json(error.message);
 		}
 	}
 
 	static async listThreeRandomTestimonials(_req: Request, res: Response) {
-        const SQL = `SELECT testimonial.id, user.name, testimonial.text, image.photo AS avatar FROM user 
+		const SQL = `SELECT testimonial.id, user.name, testimonial.text, image.photo AS avatar FROM user 
 					INNER JOIN testimonial ON testimonial.userId = user.id 
 					INNER JOIN image ON image.id = user.photoId ORDER BY RAND() LIMIT 3;`;
 		try {
 			const users: User = await manager.query(SQL);
-			return res.status(200).json(users);
+			if (users) {
+				return res.status(200).json(users);
+			} else {
+				return res.status(404).send([]);
+			}
 		} catch (error) {
 			return res.status(500).json(error.message);
 		}
